requirements.txt
wget
Pillow
numpy
opencv-python
pygame
sklearn
scipy
matplotlib
